Change in Resource.cs
- fPoints is now an int

+++ GDP/Assets/Environment/Resources/Resource.cs
@@ -12,7 +12,7 @@ public class Resource : MonoBehaviour
 	[Tooltip("The amount of horizontal offset of the resources when instantiated (The greater the number, the higher the offset)")]
 	[SerializeField] private float fMaximumOffset = 0.5f;
     [Tooltip("The amount of points for each resource node")]
-    [SerializeField] private static float fPoints = 10f;
+    [SerializeField] private static int fPoints = 10f;
 
     // Uneditable Variables
 	private Vector3 endPosition;
@@ -69,5 +69,5 @@ public class Resource : MonoBehaviour
 	}
 
     // Getter-Setter Functions
-    public static float Points { get { return fPoints; } }
+    public static int Points { get { return fPoints; } }
 }
